Split up `access_method` module into submodules
